+++ vocal_helper/__init__.py
@@ -176,4 +176,4 @@ __all__ = [
 
 __author__ = "Warith Harchaoui, Ph.D."
 __email__ = "warith@deraison.ai"
-__version__ = "0.5.2"
+__version__ = "0.6.0"